requirements.txt
aiofiles==0.7.0
aiosqlite==0.17.0
anyio==3.5.0
APScheduler==3.8.1
arrow==1.2.2
asgiref==3.5.0
bidict==0.21.4
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.3
databases==0.5.5
fastapi==0.70.1
Flask==2.0.2
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httptools==0.2.0
httpx==0.22.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jinja2-time==0.2.0
loguru==0.5.3
MarkupSafe==2.0.1
multidict==6.0.2
nb-cli==0.6.5
nonebot-adapter-onebot==2.0.0b1
nonebot2==2.0.0b1
Pillow==9.0.1
poyo==0.5.0
prompt-toolkit==3.0.26
psutil==5.9.0
pydantic==1.9.0
pyfiglet==0.8.post1
pygtrie==2.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==3.14.2
python-slugify==5.0.2
python-socketio==4.6.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.16.0
text-unidecode==1.3
tomlkit==0.7.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
waitress==2.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.0.3
yarl==1.7.2
